working fun3d tests
