+++ tests/fun3d_tests/flat_plate_mixed_fine/test_laminar_aeroelastic_mixed.py
@@ -57,11 +57,12 @@ def stdchannel_redirected(stdchannel, dest_filename):
             dest_file.close()
 
 
-@unittest.skipIf(not has_fun3d)
+@unittest.skipIf(not has_fun3d, "skipping fun3d test without fun3d")
 class TestLaminarAeroelastic(unittest.TestCase):
     FILENAME = "full_plate.txt"
     FILEPATH = os.path.join(results_folder, FILENAME)
 
+    @unittest.skip("fun3d fortran array bug")
     def test_fun3d_interface(self):
         # build a funtofem model with one body and scenario
         # build the funtofem model with one body and scenario
@@ -71,7 +72,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar", steps=500).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.ksfailure())
@@ -82,7 +83,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             # build the solvers and coupled driver
             comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(qinf=1.0e4)
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
@@ -175,7 +176,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar1", steps=200).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.lift())
@@ -184,9 +185,8 @@ class TestLaminarAeroelastic(unittest.TestCase):
         # suppress the standard output from fortran, fun3d
         with stdchannel_redirected(sys.stdout, os.devnull):
             # build the solvers and coupled driver
-            comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(qinf=1.0e4)
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
@@ -245,7 +245,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar1", steps=200).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.lift())
@@ -254,9 +254,8 @@ class TestLaminarAeroelastic(unittest.TestCase):
         # suppress the standard output from fortran, fun3d
         with stdchannel_redirected(sys.stdout, os.devnull):
             # build the solvers and coupled driver
-            comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(qinf=1.0e4)
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
