+++ tests/fun3d_tests/flat_plate_tets_coarse/test_laminar_aeroelastic.py
@@ -27,7 +27,7 @@ if has_fun3d:
 
     comm = MPI.COMM_WORLD
     results_folder = os.path.join(os.getcwd(), "results")
-    if not os.path.exists(results_folder):
+    if not os.path.exists(results_folder) and self.comm.rank == 0:
         os.mkdir(results_folder)
 
 
@@ -58,7 +58,7 @@ def stdchannel_redirected(stdchannel, dest_filename):
             dest_file.close()
 
 
-@unittest.skipIf(not has_fun3d)
+@unittest.skipIf(not has_fun3d, "skipping fun3d test without fun3d")
 class TestLaminarAeroelastic(unittest.TestCase):
     FILENAME = "full_plate.txt"
     FILEPATH = os.path.join(results_folder, FILENAME)
@@ -72,7 +72,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar", steps=500).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.ksfailure())
@@ -83,7 +83,9 @@ class TestLaminarAeroelastic(unittest.TestCase):
             # build the solvers and coupled driver
             comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(
+                qinf=1.0e4
+            )
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
@@ -176,7 +178,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar1", steps=200).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.lift())
@@ -187,7 +189,9 @@ class TestLaminarAeroelastic(unittest.TestCase):
             # build the solvers and coupled driver
             comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(
+                qinf=1.0e4
+            )
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
@@ -246,7 +250,7 @@ class TestLaminarAeroelastic(unittest.TestCase):
             lower=0.001, value=0.1, upper=2.0
         ).register_to(plate)
         plate.register_to(model)
-        test_scenario = Scenario.steady("laminar1", steps=200).set_temperature(
+        test_scenario = Scenario.steady("laminar_short", steps=200).set_temperature(
             T_ref=300.0, T_inf=300.0
         )
         test_scenario.include(Function.lift())
@@ -257,7 +261,9 @@ class TestLaminarAeroelastic(unittest.TestCase):
             # build the solvers and coupled driver
             comm = MPI.COMM_WORLD
             solvers = SolverManager(comm)
-            solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e4)
+            solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(
+                qinf=1.0e4
+            )
             solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
             # comm_manager = CommManager(comm, tacs_comm, 0, comm, 0)
             transfer_settings = TransferSettings()
