+++ tests/fun3d_tests/flat_plate_tets_coarse/test_uncoupled.py
@@ -16,13 +16,14 @@ has_fun3d = fun3d_loader is not None
 if has_fun3d:
     from pyfuntofem.interface import Fun3dInterface
 
+    comm = MPI.COMM_WORLD
     np.random.seed(1234567)
     results_folder = os.path.join(os.getcwd(), "results")
     if not os.path.exists(results_folder):
         os.mkdir(results_folder)
 
 
-@unittest.skipIf(not has_fun3d)
+@unittest.skipIf(not has_fun3d, "skipping fun3d test since you don't have fun3d")
 class TestFun3dUncoupled(unittest.TestCase):
     FILENAME = "fun3d-fake-laminar.txt"
     FILEPATH = os.path.join(results_folder, FILENAME)
@@ -46,7 +47,9 @@ class TestFun3dUncoupled(unittest.TestCase):
         # build the solvers and coupled driver
         comm = MPI.COMM_WORLD
         solvers = SolverManager(comm)
-        solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e2)
+        solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(
+            qinf=1.0e2
+        )
         solvers.structural = TestStructuralSolver(comm, model, elastic_k=1000.0)
         transfer_settings = TransferSettings()
         driver = FUNtoFEMnlbgs(
@@ -59,7 +62,8 @@ class TestFun3dUncoupled(unittest.TestCase):
         )
         self.assertTrue(max_rel_error < 1e-7)
 
-    def _laminar_aerothermal(self):
+    @unittest.skip("test aero solver thermal")
+    def test_laminar_aerothermal(self):
         # build the funtofem model with one body and scenario
         model = FUNtoFEMmodel("plate")
         plate = Body.aerothermal("plate", boundary=6)
@@ -78,7 +82,7 @@ class TestFun3dUncoupled(unittest.TestCase):
         # build the solvers and coupled driver
         comm = MPI.COMM_WORLD
         solvers = SolverManager(comm)
-        solvers.flow = Fun3dInterface(comm, model)
+        solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes")
         solvers.structural = TestStructuralSolver(comm, model, thermal_k=1.0)
         transfer_settings = TransferSettings()
         driver = FUNtoFEMnlbgs(
@@ -91,7 +95,8 @@ class TestFun3dUncoupled(unittest.TestCase):
         )
         self.assertTrue(max_rel_error < 1e-7)
 
-    def _laminar_aerothermoelastic(self):
+    @unittest.skip("test aero solver thermal")
+    def test_laminar_aerothermoelastic(self):
         # build the funtofem model with one body and scenario
         model = FUNtoFEMmodel("plate")
         plate = Body.aerothermoelastic("plate", boundary=1)
@@ -110,7 +115,9 @@ class TestFun3dUncoupled(unittest.TestCase):
         # build the solvers and coupled driver
         comm = MPI.COMM_WORLD
         solvers = SolverManager(comm)
-        solvers.flow = Fun3dInterface(comm, model).set_units(qinf=1.0e2)
+        solvers.flow = Fun3dInterface(comm, model, fun3d_dir="meshes").set_units(
+            qinf=1.0e2
+        )
         solvers.structural = TestStructuralSolver(
             comm, model, elastic_k=1000.0, thermal_k=1.0
         )
